Mode 0,1,2(port série),3,9,10(port série),11

+++ intelligentSensor/Measure.h
@@ -10,7 +10,8 @@
 struct Measure{
 
 float data=-1;
-int time=-1;
+unsigned short time=-1;
+
 };
 
 #endif
